requirements.txt
streamlit==1.23.1
tensorflow==2.13.0
numpy==1.23.5
Pillow==9.5.0
opencv-python-headless==4.7.0.72
protobuf==3.20.3